requirements.txt
stravalib>=2.0
requests>=2.31.0
python-dotenv>=1.0.0
polyline>=2.0.0

